requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
psycopg2-binary

# Retrieval / ML
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2

# Optional (for better retrieval)
faiss-cpu==1.8.0
rank-bm25==0.2.2
sentence-transformers==3.0.1

# Frontend UI
streamlit==1.37.1
requests==2.32.3

# Utilities
python-dotenv==1.0.1

